Restore subscription checks, fix cat creation navigation, and add loading illustration
- Restore real subscription checks in scan tracking and paywall bloc (testing overrides removed)
- Add timeout and error handling to RevenueCat subscription check to prevent hanging on simulator
- Fix blank screen after cat creation by using maybePop() instead of popUntilRouteWithName()
- Replace paywall loading spinner with AppLoadingWidget illustration

+++ lib/features/product/domain/entities/product_entity.dart
@@ -3,12 +3,12 @@ class ProductEntity {
   final String brand;
   final int score;
   final String imageUrl;
-  final int protein;
-  final int fat;
-  final int carbs;
-  final int fiber;
-  final int moisture;
-  final int ash;
+  final double protein;
+  final double fat;
+  final double carbs;
+  final double fiber;
+  final double moisture;
+  final double ash;
   final List<String> pros;
   final List<String> cons;
 
@@ -17,12 +17,12 @@ class ProductEntity {
     required this.brand,
     required this.score,
     required this.imageUrl,
-    this.protein = 0,
-    this.fat = 0,
-    this.carbs = 0,
-    this.fiber = 0,
-    this.moisture = 0,
-    this.ash = 0,
+    this.protein = 0.0,
+    this.fat = 0.0,
+    this.carbs = 0.0,
+    this.fiber = 0.0,
+    this.moisture = 0.0,
+    this.ash = 0.0,
     this.pros = const [],
     this.cons = const [],
   });
